requirements.txt
cv_bridge==1.13.0.post0
firebase_admin==7.1.0
matplotlib==3.5.1
numpy==1.24.4
numpy==1.21.5
openai==2.21.0
opencv_python==4.13.0.90
openwakeword==0.6.0
pandas==3.0.1
pyaudio==0.2.14
pymodbus==2.5.3
pytest==6.2.5
python-dotenv==1.2.1
scipy==1.8.0
seaborn==0.13.2
setuptools==59.6.0
sounddevice==0.5.5
torch==2.6.0+cu124
ultralytics==8.4.11
